requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0,<=4.10.0.84
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
PyYAML>=6.0
tqdm>=4.64.0